% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_allconcepts}
\alias{rx_allconcepts}
\title{Get RxNorm Concept Information}
\usage{
rx_allconcepts(tty)
}
\arguments{
\item{tty}{A list of one or more RxNorm term types.}
}
\value{
Content information
}
\description{
Get concept information for specified term types.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getAllConceptsByTTY}{RxNorm}.
}

