% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_allndcs}
\alias{rx_allndcs}
\title{Get the National Drug Codes (NDCs) for the RxNorm concept.}
\usage{
rx_allndcs(rxcui, history = 1)
}
\arguments{
\item{rxcui}{The RxNorm Concept Unique Identifier (RXCUI)}

\item{history}{(optional) if the value is 1 or not specified, all NDCs, past
or present, are returned. A value of 0 indicates only currently associated
NDCs with the concept will be returned.}
}
\value{
NDC, along with the start and end times (format:"YYYYMM")
}
\description{
The NDCs are returned in the CMS 11-digit NDC derivative form, along with
the start and end times (format:"YYYYMM") corresponding to the time interval
in which the NDC was associated with the concept.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getAllNDCs}{RxNorm}.
}

