% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_allProperties}
\alias{rx_allProperties}
\title{Get properties for a specified RxNorm concept.}
\usage{
rx_allProperties(rxcui, prop = "all")
}
\arguments{
\item{prop}{Property categories for the properties to be returned.}
}
\value{
Property name, value and category.
}
\description{
Return the properties for a specified RxNorm concept. Information returned
includes property name, value and category
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getAllProperties}{RxNorm}.
}

