% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_brands}
\alias{rx_brands}
\title{Get brands that contain all specified ingredients.}
\usage{
rx_brands(ingredientids)
}
\arguments{
\item{ingredientids}{list of (or single) ingredient RxCUIs.}
}
\value{
Brand names
}
\description{
Get the brands that contain all the specified ingredients. Note that the
brands returned may contain other ingredients in addition to those specified.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getMultiIngredBrand}{RxNorm}.
}

