% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_displaynames}
\alias{rx_displaynames}
\title{Get RxNav names}
\usage{
rx_displaynames()
}
\value{
Names of ingredients, brands, and branded packs.
}
\description{
Gets the names used by RxNav for auto completion. A large list which includes
names of ingredients, brands, and branded packs.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getDisplayTerms}{RxNorm}.
}

