% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_drugs}
\alias{rx_drugs}
\title{Get Drug Information}
\usage{
rx_drugs(drugName)
}
\arguments{
\item{drugName}{An ingredient, brand, clinical dose form, branded dose form, clinical drug component or branded drug component name}
}
\value{
Some results.
}
\description{
Get the drug products associated with a specified name. The name can be an ingredient, brand name, clinical dose form, branded dose form, clinical drug component, or branded drug component. See \href{http://mor.nlm.nih.gov/download/rxnav/RxNormAPIREST.html#uLink=RxNorm_REST_getDrugs}{RxNorm}.
}

