% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_filter}
\alias{rx_filter}
\title{Determine if a property exists for a concept}
\usage{
rx_filter(rxcui, propName, propValues = "IN")
}
\arguments{
\item{rxcui}{The RxNorm Concept Unique Identifier (RXCUI)}

\item{propName}{Property name}

\item{propValues}{(optional) Property value.}
}
\value{
RxCUI if the property name matches.
}
\description{
Determine if a property exists for a concept and (optionally) matches the
specified property value. Returns the RxCUI if the property name matches.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_filterByProperty}{RxNorm}.
}

