% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_ndcproperties}
\alias{rx_ndcproperties}
\title{Get the National Drug Code (NDC) properties.}
\usage{
rx_ndcproperties(id)
}
\arguments{
\item{id}{Valid identifiers are NDC 2 segment, NDC 3 segment, NDC11, RxCUI,
and SPL_SET_ID}
}
\value{
NDC properties, NDC, or structured product label.
}
\description{
This function returns the NDC properties for an RxNorm concept, an NDC or a
structured product label. Only NDC properties of active NDCs curated by
RxNorm are returned.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getNDCProperties}{RxNorm}.
}

