% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_ndcs}
\alias{rx_ndcs}
\title{Get the active National Drug Codes (NDCs) for the RxNorm concept.}
\usage{
rx_ndcs(rxcui)
}
\value{
National Drug Codes (NDCs) for the RxNorm concept.
}
\description{
Return the active National Drug Codes (NDCs) for the RxNorm concept. Active
NDCs are those NDCs present in the current version of RxNorm. Only the NDCs
curated by RxNorm (i.e., with SAB = RXNORM) are returned by this function.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getNDCs}{RxNorm}.
}

