% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_ndcstatus}
\alias{rx_ndcstatus}
\title{Get the status for a National Drug Code (NDC).}
\usage{
rx_ndcstatus(ndc, start = NULL, end = NULL, history = 0)
}
\arguments{
\item{ndc}{National Drug Code (NDC).}

\item{start}{(optional) Starting date used in conjunction with an end date to
provide an interval to return history data. If specified, an end date must
also be specified.}

\item{end}{(optional) Ending date used in conjuction with a start date to
provide an interval to return history data.If specified, a start date must
also be specified.}

\item{history}{(optional, default:0) 0 = include all history. 1 = include
only lastest entry}
}
\value{
NDC status and history data for the NDC.
}
\description{
This resource returns the NDC status along with history data for the NDC.
Possible status values returned are:
}
\details{
Active: The NDC is contained in the current data set.

Obsolete: The NDC previously existed in RxNorm, but is no longer recognized
as active.

Alien: The NDC exists in the current data set, but only in a source
vocabulary other than RxNorm.

Unknown: The NDC is unknown or invalid.

See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getNDCStatus}{RxNorm}.
}

