% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_properties}
\alias{rx_properties}
\title{Get the RxNorm concept properties.}
\usage{
rx_properties(rxcui)
}
\arguments{
\item{rxcui}{The RxNorm Concept Unique Identifier (RXCUI)}
}
\value{
Concept properties.
}
\description{
Return the RxNorm concept properties: Concept name, Concept identifier
(RxCUI), Synonym, RxNorm term type, Language of the term, UMLS CUI,
Suppress flag
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getRxConceptProperties}{RxNorm}.
}

