% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_property}
\alias{rx_property}
\title{Get the property values associated with the property name.}
\usage{
rx_property(rxcui, propName)
}
\arguments{
\item{rxcui}{The RxNorm Concept Unique Identifier (RXCUI)}

\item{propName}{the property name. See {\link{rx_propnames}} for the list of
valid property names}
}
\value{
Property values associated with the property name.
}
\description{
This function returns the property values associated with the property name.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getRxProperty}{RxNorm}.
}

