% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_related_rela}
\alias{rx_related_rela}
\title{Get the related RxNorm identifiers of an RxNorm concept by relational attribute.}
\usage{
rx_related_rela(rxcui, rela)
}
\arguments{
\item{rela}{A list of the relationship attribute names such as
"tradename_of", "has_form", "isa", etc. This field is required. See the
{\link{rx_relatypes}} example for the valid relationship attributes.}
}
\value{
Related RxNorm identifiers.
}
\description{
This function returns the related RxNorm identifiers of an RxNorm concept
specified by a relational attribute.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getRelatedByRelationship}{RxNorm}.
}

