% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_related_tty}
\alias{rx_related_tty}
\title{Get the related RxNorm identifiers of an RxNorm concept by term types.}
\usage{
rx_related_tty(rxcui, tty)
}
\arguments{
\item{tty}{- a list of one or more RxNorm term types. This field is required.
See the {\link{rx_termtypes}} example for the valid term types.}
}
\value{
Related RxNorm identifiers.
}
\description{
This function returns the related RxNorm identifiers of an RxNorm concept
specified by one or more term types.
See \href{https://rxnav.nlm.nih.gov/RxNavViews.html#label:appendix}{default paths}
for the paths traveled to get concepts for each term type.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getRelatedByType}{RxNorm}.
}

