% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_relatypes}
\alias{rx_relatypes}
\title{Get relationship names.}
\usage{
rx_relatypes()
}
\value{
Relationship names.
}
\description{
Return the valid relationship names.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getRelaTypes}{RxNorm}.
}

