% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_rxcui_idtype}
\alias{rx_rxcui_idtype}
\title{Get RxCUIs from another vocabulary identifier.}
\usage{
rx_rxcui_idtype(idtype, id, allsrc = 0)
}
\arguments{
\item{idtype}{Identifier type. Valid types are AMPID, ANDA, ATC, GCN_SEQNO,
GFC, GPPC, HIC_SEQN, MESH, MMSL_CODE, NDA, NDC, NUI, SNOMEDCT, SPL_SET_ID,
UMLSCUI, UNII_CODE, and VUID}

\item{id}{Identifier.}

\item{allsrc}{Flag indicating whether to consider all identifiers in the
RxNorm data set (=1) or only identifiers which are related synonyms to an
RxNorm vocabulary term (=0).}
}
\value{
RxCUIs
}
\description{
Search for an identifier from another vocabulary and return the RxCUIs of any
concepts which have an RxNorm term as a synonym or have that identifier as an
attribute.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_findRxcuiById}{RxNorm}.
}

