% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_rxcui_name}
\alias{rx_rxcui_name}
\title{Get RxCUIs for a ame in RxNorm.}
\usage{
rx_rxcui_name(name, srclist = NULL, allsrc = 0, search = 0)
}
\arguments{
\item{name}{Search string.}

\item{srclist}{List of the source vocabulary names to use. Only used when
allsrc is set to 1. See the sourcetypes() example for the valid source
vocabularies. If not specified, all sources are used.}

\item{allsrc}{Flag indicating whether to consider all terms in the RxNorm
data set (=1) or only terms which are either from the RxNorm vocabulary or
related synonyms from other vocabularies (=0).}

\item{search}{Number indicating the type of search to be performed. If set to
0 (the default), exact match search will be performed. If set to 1, a
normalized string search will be done. When search = 2, then an exact match
search will be done, and if no results are found, a normalized search will be
done.}
}
\value{
RxCUIs
}
\description{
Search for a name in the RxNorm data set and return the RxCUIs of any
concepts which have that name as an RxNorm term or as a synonym of an RxNorm
term. Search options include exact match searching or normalized string
search.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_findRxcuiByString}{RxNorm}.
}

