% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_spellingsuggestions}
\alias{rx_spellingsuggestions}
\title{Get spelling suggestions for a given term.}
\usage{
rx_spellingsuggestions(name)
}
\arguments{
\item{name}{Name for which spelling suggestions are to be generated.}
}
\value{
RxNorm terms.
}
\description{
Return spelling suggestions for a given term. The suggestions are RxNorm
terms contained in the current version, listed in decreasing order of
closeness to the original phrase.
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getSpellingSuggestions}{RxNorm}.
}

