% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_status}
\alias{rx_status}
\title{Get the status for a concept.}
\usage{
rx_status(rxcui)
}
\arguments{
\item{rxcui}{The RxNorm Concept Unique Identifier (RXCUI)}
}
\value{
Concept status.
}
\description{
Return the status for a concept. Possible values are:
Active (The concept is in the current data set and has a non-obsolete term
from the RxNorm vocabulary)
Alien (The concept exists in the current data set, but contains only terms
from vocabularies other than RxNorm.)
Remapped (The concept has been remapped to one or more concepts in the
current data set.)
Retired (The concept no longer exists in the current data set, or contains
only obsolete terms.)
Unknown (The concept identifier is invalid.)
}
\details{
See \href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getRxcuiStatus}{RxNorm}.
}

