% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_termtypes}
\alias{rx_termtypes}
\title{Get RxNorm Term Types}
\usage{
rx_termtypes()
}
\value{
Term type information.
}
\description{
Get the valid term types in the RxNorm data set. See \href{http://mor.nlm.nih.gov/download/rxnav/RxNormAPIREST.html#uLink=RxNorm_REST_getTermTypes}{RxNorm}.
}

