% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_version}
\alias{rx_version}
\title{Get Version of API}
\usage{
rx_version()
}
\value{
Version information.
}
\description{
Get the version of the RxNorm data set. See \href{http://mor.nlm.nih.gov/download/rxnav/RxNormAPIREST.html#uLink=RxNorm_REST_getRxNormVersion}{RxNorm}.
}

